Move utility yaml extensions into a separate module for simplicity
